+++ rubato/struct/__init__.py
@@ -1,8 +1,8 @@
 """
 The classes module holds all of the high-dependency modules relating to game objects.
 """
-from .sprite import Sprite
 from .surface import Surface
+from .sprite import Sprite
 from .gameobject import *
 from .group import Group
 from .scene import Scene
